Fix error

+++ MuscleRecord/ViewModel/ViewModel.swift
@@ -53,53 +53,57 @@ class ViewModel: ObservableObject {
     
     func updateEvent(event: Event, newName: String) {
         let db = Firestore.firestore()
-        let userID = Auth.auth().currentUser!.uid
-        db.collection("users").document(userID).collection("events").document(event.id).updateData(["name": newName]) { error in
-            if let error = error {
-                print("updateEvent中のエラー: \(error)")
-            } else {
-                print("updateEvent成功")
+        if let userID = Auth.auth().currentUser?.uid {
+            db.collection("users").document(userID).collection("events").document(event.id).updateData(["name": newName]) { error in
+                if let error = error {
+                    print("updateEvent中のエラー: \(error)")
+                } else {
+                    print("updateEvent成功")
+                }
             }
         }
     }
     
     func deleteEvent(event: Event) {
         let db = Firestore.firestore()
-        let userID = Auth.auth().currentUser!.uid
-        db.collection("users").document(userID).collection("events").document(event.id).delete() { error in
-            if let error = error {
-                print("deleteEvent中のエラー: \(error)")
-            } else {
-                print("deleteEvent成功")
+        if let userID = Auth.auth().currentUser?.uid {
+            db.collection("users").document(userID).collection("events").document(event.id).delete() { error in
+                if let error = error {
+                    print("deleteEvent中のエラー: \(error)")
+                } else {
+                    print("deleteEvent成功")
+                }
             }
         }
     }
     
     func addEvent(_ name: String) {
         let db = Firestore.firestore()
-        let userID = Auth.auth().currentUser!.uid
-        db.collection("users").document(userID).collection("events").addDocument(data: ["name": name, "latestWeight": 0.0, "latestRep": 0, "latestDate": Date(timeInterval: -60*60*24, since: .now)]) { error in
-            if let error = error {
-                print("addEvent中のエラー: \(error)")
-            } else {
-                print("addEvent成功")
+        if let userID = Auth.auth().currentUser?.uid {
+            db.collection("users").document(userID).collection("events").addDocument(data: ["name": name, "latestWeight": 0.0, "latestRep": 0, "latestDate": Date(timeInterval: -60*60*24, since: .now)]) { error in
+                if let error = error {
+                    print("addEvent中のエラー: \(error)")
+                } else {
+                    print("addEvent成功")
+                }
             }
         }
     }
     
     func getEvent() {
         let db = Firestore.firestore()
-        let userID = Auth.auth().currentUser!.uid
-        db.collection("users").document(userID).collection("events").getDocuments { snapshot, error in
-            if let snapshot = snapshot{
-                if let error = error {
-                    print("getEvent中のエラー: \(error)")
-                } else {
-                    print("getEvent成功")
-                    DispatchQueue.main.async {
-                        self.events = snapshot.documents.map { d in
-                            let timeStamp = d["latestDate"] as! Timestamp
-                            return Event(id: d.documentID, name: d["name"] as? String ?? "", latestWeight: d["latestWeight"] as? Float ?? 0.0, latestRep: d["latestRep"] as? Int ?? 0, latestDate: timeStamp.dateValue())
+        if let userID = Auth.auth().currentUser?.uid {
+            db.collection("users").document(userID).collection("events").getDocuments { snapshot, error in
+                if let snapshot = snapshot{
+                    if let error = error {
+                        print("getEvent中のエラー: \(error)")
+                    } else {
+                        print("getEvent成功")
+                        DispatchQueue.main.async {
+                            self.events = snapshot.documents.map { d in
+                                let timeStamp = d["latestDate"] as! Timestamp
+                                return Event(id: d.documentID, name: d["name"] as? String ?? "", latestWeight: d["latestWeight"] as? Float ?? 0.0, latestRep: d["latestRep"] as? Int ?? 0, latestDate: timeStamp.dateValue())
+                            }
                         }
                     }
                 }
@@ -109,132 +113,136 @@ class ViewModel: ObservableObject {
     
     func getRecord(event: Event) {
         let db = Firestore.firestore()
-        let userID = Auth.auth().currentUser!.uid
-        db.collection("users").document(userID).collection("events").document(event.id).collection("records").order(by: "date").getDocuments { (snapshot, error) in
-            if let snapshot = snapshot{
-                if let error = error {
-                    print("getRecord中のエラー: \(error)")
-                } else {
-                    print("getRecord成功")
-                    DispatchQueue.main.async {
-                        snapshot.documents.forEach { d in
-                            self.latestID = d.documentID
-                            let timeStamp: Timestamp = d["date"] as? Timestamp ?? Timestamp()
-                            let date = timeStamp.dateValue()
-                            if let oldRecord = self.oldRecord {
-                                var dateDifference = (Calendar.current.dateComponents([.day], from: oldRecord.date, to: timeStamp.dateValue())).day! - 1
-                                while dateDifference > 0 {
-                                    self.records.append(Record(id: UUID().uuidString, date: Date(timeInterval: TimeInterval(-60*60*24*dateDifference), since: date), weight: oldRecord.weight, rep: oldRecord.rep, dummy: true))
-                                    dateDifference -= 1
-                                }
-                            }
-                            let record = Record(id: d.documentID, date: date, weight: d["weight"] as? Float ?? 0, rep: d["rep"] as? Int ?? 0, dummy: false)
-                            self.oldRecord = record
-                            self.records.append(record)
-                        }
-                        
-                        //取得した記録を元に3日、9日、27日平均に変換した配列を作る
-                        let periodArray = [3,9,27]
-                        for period in periodArray {
-                            var totalDate = 0
-                            var totalWeight: Float = 0
-                            var totalRep = 0
-                            let fraction = self.records.count % period
-                            var created = false
-                            self.records.forEach { record in
-                                totalWeight += record.weight
-                                totalRep += record.rep
-                                totalDate += 1
-                                if fraction != 0 && fraction == totalDate && created == false {
-                                    let recordID = UUID().uuidString
-                                    self.records3.append(Record(id: recordID, date: record.date, weight: totalWeight/Float(totalDate), rep: totalRep/totalDate, dummy: false))
-                                    self.latestID3 = recordID
-                                    created = true
-                                    totalWeight = 0
-                                    totalRep = 0
-                                    totalDate = 0
-                                }
-                                if totalDate == period {
-                                    let recordID = UUID().uuidString
-                                    if period == 3 {
-                                        self.records3.append(Record(id: recordID, date: record.date, weight: totalWeight/Float(period), rep: totalRep/period, dummy: false))
-                                        self.latestID3 = recordID
-                                    } else if period == 9 {
-                                        self.records9.append(Record(id: recordID, date: record.date, weight: totalWeight/Float(period), rep: totalRep/period, dummy: false))
-                                        self.latestID9 = recordID
-                                    } else if period == 27 {
-                                        self.records27.append(Record(id: recordID, date: record.date, weight: totalWeight/Float(period), rep: totalRep/period, dummy: false))
-                                        self.latestID27 = recordID
+        if let userID = Auth.auth().currentUser?.uid {
+            db.collection("users").document(userID).collection("events").document(event.id).collection("records").order(by: "date").getDocuments { (snapshot, error) in
+                if let snapshot = snapshot{
+                    if let error = error {
+                        print("getRecord中のエラー: \(error)")
+                    } else {
+                        print("getRecord成功")
+                        DispatchQueue.main.async {
+                            snapshot.documents.forEach { d in
+                                self.latestID = d.documentID
+                                let timeStamp: Timestamp = d["date"] as? Timestamp ?? Timestamp()
+                                let date = timeStamp.dateValue()
+                                if let oldRecord = self.oldRecord {
+                                    var dateDifference = (Calendar.current.dateComponents([.day], from: oldRecord.date, to: timeStamp.dateValue())).day! - 1
+                                    while dateDifference > 0 {
+                                        self.records.append(Record(id: UUID().uuidString, date: Date(timeInterval: TimeInterval(-60*60*24*dateDifference), since: date), weight: oldRecord.weight, rep: oldRecord.rep, dummy: true))
+                                        dateDifference -= 1
                                     }
-                                    totalWeight = 0
-                                    totalRep = 0
-                                    totalDate = 0
+                                }
+                                let record = Record(id: d.documentID, date: date, weight: d["weight"] as? Float ?? 0, rep: d["rep"] as? Int ?? 0, dummy: false)
+                                self.oldRecord = record
+                                self.records.append(record)
+                            }
+                            
+                            //取得した記録を元に3日、9日、27日平均に変換した配列を作る
+                            let periodArray = [3,9,27]
+                            for period in periodArray {
+                                var totalDate = 0
+                                var totalWeight: Float = 0
+                                var totalRep = 0
+                                let fraction = self.records.count % period
+                                var created = false
+                                self.records.forEach { record in
+                                    totalWeight += record.weight
+                                    totalRep += record.rep
+                                    totalDate += 1
+                                    if fraction != 0 && fraction == totalDate && created == false {
+                                        let recordID = UUID().uuidString
+                                        self.records3.append(Record(id: recordID, date: record.date, weight: totalWeight/Float(totalDate), rep: totalRep/totalDate, dummy: false))
+                                        self.latestID3 = recordID
+                                        created = true
+                                        totalWeight = 0
+                                        totalRep = 0
+                                        totalDate = 0
+                                    }
+                                    if totalDate == period {
+                                        let recordID = UUID().uuidString
+                                        if period == 3 {
+                                            self.records3.append(Record(id: recordID, date: record.date, weight: totalWeight/Float(period), rep: totalRep/period, dummy: false))
+                                            self.latestID3 = recordID
+                                        } else if period == 9 {
+                                            self.records9.append(Record(id: recordID, date: record.date, weight: totalWeight/Float(period), rep: totalRep/period, dummy: false))
+                                            self.latestID9 = recordID
+                                        } else if period == 27 {
+                                            self.records27.append(Record(id: recordID, date: record.date, weight: totalWeight/Float(period), rep: totalRep/period, dummy: false))
+                                            self.latestID27 = recordID
+                                        }
+                                        totalWeight = 0
+                                        totalRep = 0
+                                        totalDate = 0
+                                    }
+                                }
+                            }
+                            
+                            //グラフの目盛り幅を決めるために重量の最大値を計算
+                            snapshot.documents.forEach { d in
+                                if self.maxWeight < d["weight"] as? Float ?? 0.0 {
+                                    self.maxWeight = d["weight"] as? Float ?? 0.0
                                 }
                             }
                         }
                         
-                        //グラフの目盛り幅を決めるために重量の最大値を計算
-                        snapshot.documents.forEach { d in
-                            if self.maxWeight < d["weight"] as? Float ?? 0.0 {
-                                self.maxWeight = d["weight"] as? Float ?? 0.0
-                            }
-                        }
                     }
-                    
                 }
             }
+
         }
     }
     
     func addRecord(event: Event, weight: Float, rep: Int) {
         let db = Firestore.firestore()
-        let userID = Auth.auth().currentUser!.uid
-        db.collection("users").document(userID).collection("events").document(event.id).setData(["name": event.name, "latestWeight": weight, "latestRep": rep, "latestDate": Date()]) { error in
-            if let error = error {
-                print("addRecord(event)中のエラー: \(error)")
-            } else {
-                print("addRecord(event)成功")
+        if let userID = Auth.auth().currentUser?.uid {
+            db.collection("users").document(userID).collection("events").document(event.id).setData(["name": event.name, "latestWeight": weight, "latestRep": rep, "latestDate": Date()]) { error in
+                if let error = error {
+                    print("addRecord(event)中のエラー: \(error)")
+                } else {
+                    print("addRecord(event)成功")
+                }
             }
-        }
-        db.collection("users").document(userID).collection("events").document(event.id).collection("records").addDocument(data: ["date": Date(), "weight": weight, "rep": rep]) { error in
-            if let error = error {
-                print("addRecord(record)中のエラー: \(error)")
-            } else {
-                print("addRecord(record)成功")
+            db.collection("users").document(userID).collection("events").document(event.id).collection("records").addDocument(data: ["date": Date(), "weight": weight, "rep": rep]) { error in
+                if let error = error {
+                    print("addRecord(record)中のエラー: \(error)")
+                } else {
+                    print("addRecord(record)成功")
+                }
             }
         }
     }
     
     func updateRecord(event: Event, weight: Float, rep: Int) {
         let db = Firestore.firestore()
-        let userID = Auth.auth().currentUser!.uid
-        db.collection("users").document(userID).collection("events").document(event.id).collection("records").order(by: "date", descending: true).limit(to: 1).getDocuments { snapshot, error in
-            if let snapshot = snapshot {
-                if let error = error {
-                    print("updateRecord(delete)中のエラー: \(error)")
-                } else {
-                    print("updateRecord(delete)成功")
-                    snapshot.documents.forEach { d in
-                        let latestRecordID = d.documentID
-                        db.collection("users").document(userID).collection("events").document(event.id).collection("records").document(latestRecordID).delete()
+        if let userID = Auth.auth().currentUser?.uid {
+            db.collection("users").document(userID).collection("events").document(event.id).collection("records").order(by: "date", descending: true).limit(to: 1).getDocuments { snapshot, error in
+                if let snapshot = snapshot {
+                    if let error = error {
+                        print("updateRecord(delete)中のエラー: \(error)")
+                    } else {
+                        print("updateRecord(delete)成功")
+                        snapshot.documents.forEach { d in
+                            let latestRecordID = d.documentID
+                            db.collection("users").document(userID).collection("events").document(event.id).collection("records").document(latestRecordID).delete()
+                        }
                     }
                 }
             }
-        }
-        
-        db.collection("users").document(userID).collection("events").document(event.id).setData(["latestWeight": weight, "latestRep": rep,"latestDate": Date()]) { error in
-            if let error = error {
-                print("updateRecord(event)中のエラー: \(error)")
-            } else {
-                print("updateRecord(event)成功")
+            
+            db.collection("users").document(userID).collection("events").document(event.id).setData(["latestWeight": weight, "latestRep": rep,"latestDate": Date()]) { error in
+                if let error = error {
+                    print("updateRecord(event)中のエラー: \(error)")
+                } else {
+                    print("updateRecord(event)成功")
+                }
             }
-        }
-        
-        db.collection("users").document(userID).collection("events").document(event.id).collection("records").addDocument(data: ["date": Date(), "weight": weight, "rep": rep]) { error in
-            if let error = error {
-                print("updateRecord(record)中のエラー: \(error)")
-            } else {
-                print("updateRecord(record)成功")
+            
+            db.collection("users").document(userID).collection("events").document(event.id).collection("records").addDocument(data: ["date": Date(), "weight": weight, "rep": rep]) { error in
+                if let error = error {
+                    print("updateRecord(record)中のエラー: \(error)")
+                } else {
+                    print("updateRecord(record)成功")
+                }
             }
         }
     }
